Sync with the private repository
Deal with acquiring 0x0 buffer request.
Prepend new buffer for give the mouse event to newly created buffer.
Because the mouse event will be feed to the first matched buffer info.
(using pkgname & id)
so the previous code is not able to give the mouse event to lately created buffer info.
it just pass the mouse event to the first created buffer.
but the first created buffer is not usable.
This patch will change the order of buffer info in its list.
So the mouse event will be feed to the last buffer info.
and the first buffer info will not be used anymore.

Add log for debugging

Update accessibility API.
To send the accessiblity result status to the viewer.
Update VERSION

Add update begin/end. (passive/active) update mode

Update mode event handler added.
And the recreate event updated.
To recover the instance properly.
Update mode & scroll lock information is added.

Change-Id: If6a5c7c03b9230212217e05c9bf9b48c130d97ae